remove abstract graphql type again
 - push getName() implementation and property down
 - modified types return null for getName
 - equals/hashcode pushed down as well, modified types delegate to wrappedType
