+++ src/main/java/graphql/schema/GraphQLList.java
@@ -5,12 +5,11 @@ import java.util.Map;
 
 import static graphql.Assert.assertNotNull;
 
-public class GraphQLList extends AbstractGraphQLType implements GraphQLInputType, GraphQLOutputType, GraphQLModifiedType, GraphQLNullableType {
+public class GraphQLList implements GraphQLInputType, GraphQLOutputType, GraphQLModifiedType, GraphQLNullableType {
 
     private GraphQLType wrappedType;
 
     public GraphQLList(GraphQLType wrappedType) {
-        super("");
         assertNotNull(wrappedType, "wrappedType can't be null");
         this.wrappedType = wrappedType;
     }
@@ -40,4 +39,7 @@ public class GraphQLList extends AbstractGraphQLType implements GraphQLInputType
         return wrappedType != null ? wrappedType.hashCode() : 0;
     }
 
+    public String getName() {
+        return null;
+    }
 }
