+++ src/main/java/graphql/schema/GraphQLNonNull.java
@@ -5,12 +5,11 @@ import java.util.Map;
 
 import static graphql.Assert.assertNotNull;
 
-public class GraphQLNonNull extends AbstractGraphQLType implements GraphQLInputType, GraphQLOutputType, GraphQLModifiedType {
+public class GraphQLNonNull implements GraphQLInputType, GraphQLOutputType, GraphQLModifiedType {
 
     private  GraphQLType wrappedType;
 
     public GraphQLNonNull(GraphQLType wrappedType) {
-        super("");
         assertNotNull(wrappedType, "wrappedType can't be null");
         this.wrappedType = wrappedType;
     }
