+++ src/main/java/graphql/schema/GraphQLScalarType.java
@@ -3,15 +3,16 @@ package graphql.schema;
 
 import static graphql.Assert.assertNotNull;
 
-public class GraphQLScalarType extends AbstractGraphQLType implements GraphQLInputType, GraphQLOutputType, GraphQLUnmodifiedType,GraphQLNullableType {
+public class GraphQLScalarType implements GraphQLInputType, GraphQLOutputType, GraphQLUnmodifiedType,GraphQLNullableType {
 
+    private final String name;
     private final String description;
     private final Coercing coercing;
 
-
     public GraphQLScalarType(String name, String description, Coercing coercing) {
-        super(name);
+        assertNotNull(name, "name can't null");
         assertNotNull(coercing, "coercing can't be null");
+        this.name = name;
         this.description = description;
         this.coercing = coercing;
     }
@@ -33,5 +34,25 @@ public class GraphQLScalarType extends AbstractGraphQLType implements GraphQLInp
                 ", description='" + description + '\'' +
                 ", coercing=" + coercing +
                 '}';
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+
+        GraphQLType that = (GraphQLType) o;
+
+        return getName().equals(that.getName());
+
+    }
+
+    @Override
+    public int hashCode() {
+        return name.hashCode();
     }
 }
